cli: rearrange versioning commands
Reparent everything having to do with managing versions under the
`esc env version` command and rearrange the command structure somewhat:

- `esc env version`: decribes a revision or tag
- `esc env version history`: show revision history
- `esc env version tag`: create or update a tagged revision
- `esc env version tag ls`: list tagged revisions
- `esc env version tag rm`: remove a tagged revision
