+++ cmd/esc/cli/env_init.go
@@ -39,7 +39,7 @@ func newEnvInitCmd(env *envCommand) *cobra.Command {
 				return err
 			}
 			if version != "" {
-				return fmt.Errorf("the init command does not accept revisions or tags")
+				return fmt.Errorf("the init command does not accept versions")
 			}
 			_ = args
