+++ cmd/esc/cli/testdata/env-version-tag.yaml
@@ -1,16 +1,16 @@
 run: |
   esc env version tag test@stable 2
-  esc env version show test@stable --utc
+  esc env version test@stable --utc
   esc env version tag test@stable 3
-  esc env version show test@stable --utc
+  esc env version test@stable --utc
   esc env version tag test@stable
-  esc env version show test@stable --utc
-  esc env version show test@latest --utc
-  esc env version ls test --utc
+  esc env version test@stable --utc
+  esc env version test@latest --utc
+  esc env version tag ls test --utc
   esc env version tag test@initial 1
-  esc env version ls test --utc
-  esc env version rm test@initial
-  esc env version ls test --utc
+  esc env version tag ls test --utc
+  esc env version tag rm test@initial
+  esc env version tag ls test --utc
 environments:
   test-user/a: {}
   test-user/b: {}
@@ -40,25 +40,25 @@ environments:
                 ciphertext: ZXNjeAAAAAHz5ePy5fTB4+Pl8/PL5fnJxPD7
 stdout: |+
   > esc env version tag test@stable 2
-  > esc env version show test@stable --utc
+  > esc env version test@stable --utc
   stable
   Revision 2
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
   > esc env version tag test@stable 3
-  > esc env version show test@stable --utc
+  > esc env version test@stable --utc
   stable
   Revision 3
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
   > esc env version tag test@stable
-  > esc env version show test@stable --utc
+  > esc env version test@stable --utc
   stable
   Revision 4
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
-  > esc env version show test@latest --utc
+  > esc env version test@latest --utc
   latest
   Revision 4
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
-  > esc env version ls test --utc
+  > esc env version tag ls test --utc
   latest
   Revision 4
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
@@ -68,7 +68,7 @@ stdout: |+
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
 
   > esc env version tag test@initial 1
-  > esc env version ls test --utc
+  > esc env version tag ls test --utc
   initial
   Revision 1
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
@@ -81,8 +81,8 @@ stdout: |+
   Revision 4
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
 
-  > esc env version rm test@initial
-  > esc env version ls test --utc
+  > esc env version tag rm test@initial
+  > esc env version tag ls test --utc
   latest
   Revision 4
   Last updated at 0001-01-01 00:00:00 +0000 UTC by <unknown>
@@ -93,14 +93,14 @@ stdout: |+
 
 stderr: |
   > esc env version tag test@stable 2
-  > esc env version show test@stable --utc
+  > esc env version test@stable --utc
   > esc env version tag test@stable 3
-  > esc env version show test@stable --utc
+  > esc env version test@stable --utc
   > esc env version tag test@stable
-  > esc env version show test@stable --utc
-  > esc env version show test@latest --utc
-  > esc env version ls test --utc
+  > esc env version test@stable --utc
+  > esc env version test@latest --utc
+  > esc env version tag ls test --utc
   > esc env version tag test@initial 1
-  > esc env version ls test --utc
-  > esc env version rm test@initial
-  > esc env version ls test --utc
+  > esc env version tag ls test --utc
+  > esc env version tag rm test@initial
+  > esc env version tag ls test --utc
